requirements.txt
streamlit
pandas
scikit-learn==1.6.1
joblib
openpyxl  # Needed to read .xlsx files
numpy
requests
python-dotenv

